A lot of new features

+++ app/src/main/java/com/denspark/strelets/cinematrix/utils/DimensionUtils.java
@@ -19,7 +19,7 @@ public class DimensionUtils {
         return (int) (px / pixelsPerOneDp);
     }
 
-    public static int dpToPx(Context context, int dp) {
+    public static int dpToPx(Context context, float dp) {
         if (!isInitialised) {
             initialise(context);
         }
